+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 <!-- _No changes yet._ -->
 
 - fix: enhance YabaiService with JSON cleanup and filtering for spaces, windows, and displays
+- refactor: enhance signal handling in YabaiService
 - feat: add microphone control functionality in SystemInfoService
 - feat: enhance MicWidget with popover functionality for microphone control